+++ CHANGELOG.md
@@ -41,7 +41,7 @@ Performance:
 * Fiber-local variables are much faster now by using less synchronization.
 * Improved the performance of the exceptional case of `String#chr` (#2318, @chrisseaton).
 * Improved the performance of `IO#read_nonblock` when no data is available to be read.
-* Faster solution for lazy integer length (#2365, @chrisseaton).
+* Faster solution for lazy integer length (#2365, @lemire, @chrisseaton).
 
 Changes: